// src/types.ts
export type Role = 'user' | 'assistant'
// ...
export type User = {
    /**
     * ID of the user
     */
    id: string

    /**
     * Name of the user
     */
    name: string

    /**
     * Email of the user
     */
    email?: string

    /**
     * Image of the user
     */
    image: string

    /**
     * Picture of the user
     */
    picture: string

    /**
     * Groups the user is in
     */
    groups: string[]

    /**
     * Features the user is in
     */
    features: string[]
}
// ...
export type ConversationJSONBody = {
    /**
     * The action to take
     */
    action: string

    /**
     * The ID of the conversation
     */
    conversation_id?: string

    /**
     * Prompts to provide
     */
    messages: Prompt[]

    /**
     * The model to use
     */
    model: string

    /**
     * The parent message ID
     */
    parent_message_id: string
}

export type Prompt = {
    /**
     * The content of the prompt
     */
    content: PromptContent

    /**
     * The ID of the prompt
     */
    id: string

    /**
     * The role played in the prompt
     */
    role: Role
}

export type PromptContent = {
    /**
     * The content type of the prompt
     */
    content_type: ContentType

    /**
     * The parts to the prompt
     */
    parts: string[]
}
// ...
export type ConversationResponseEvent = {
    message?: Message
    conversation_id?: string
    error?: string | null
}

export type Message = {
    id: string
    content: MessageContent
    role: string
    user: string | null
    create_time: string | null
    update_time: string | null
    end_turn: null
    weight: number
    recipient: string
    metadata: MessageMetadata
}

export type MessageContent = {
    content_type: string
    parts: string[]
}

export type MessageMetadata = any
export type MessageActionType = 'next' | 'variant'

export type SendMessageOptions = {
    conversationId?: string
    parentMessageId?: string
    messageId?: string
    action?: MessageActionType
    timeoutMs?: number
    onProgress?: (partialResponse: string) => void
    onConversationResponse?: (response: ConversationResponseEvent) => void
    abortSignal?: AbortSignal
}

export type SendConversationMessageOptions = Omit<
    SendMessageOptions,
    'conversationId' | 'parentMessageId'
    >

export class ChatGPTError extends Error {
    statusCode?: number
    statusText?: string
    response?: AxiosResponse
    originalError?: Error
}